backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
langchain>=0.3.27
langchain-google-genai>=2.0.10
langchain-community>=0.3.28
google-generativeai>=0.8.5
google-ai-generativelanguage>=0.6.15
pydantic>=2.11.7
python-dotenv>=1.0.1
numpy>=1.24.0,<2.0.0